# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Geelink Data Ingestion System
  version: 1.1.0
paths:
...
  /ai-vocabulary/samples/generate:
    post:
      tags:
      - AI Vocabulary
      summary: Generate frozen base sample
      operationId: generate_sample_ai_vocabulary_samples_generate_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateSampleRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SampleVersionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /ai-vocabulary/samples/versions:
    get:
      tags:
      - AI Vocabulary
      summary: List sample versions
      operationId: list_sample_versions_ai_vocabulary_samples_versions_get
      parameters:
      - name: dataset
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Dataset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SampleVersionResponse'
                title: Response List Sample Versions Ai Vocabulary Samples Versions
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /ai-vocabulary/samples/versions/{sample_version_id}:
    get:
      tags:
      - AI Vocabulary
      summary: Get sample version
      operationId: get_sample_version_ai_vocabulary_samples_versions__sample_version_id__get
      parameters:
      - name: sample_version_id
        in: path
        required: true
        schema:
          type: string
          title: Sample Version Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SampleVersionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /ai-vocabulary/samples/versions/{sample_version_id}/items:
    get:
      tags:
      - AI Vocabulary
      summary: List sample items
      operationId: list_sample_items_ai_vocabulary_samples_versions__sample_version_id__items_get
      parameters:
      - name: sample_version_id
        in: path
        required: true
        schema:
          type: string
          title: Sample Version Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SampleItemResponse'
                title: Response List Sample Items Ai Vocabulary Samples Versions  Sample
                  Version Id  Items Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /ai-vocabulary/runs:
    post:
      tags:
      - AI Vocabulary
      summary: Create AI vocabulary extraction run
      operationId: create_run_ai_vocabulary_runs_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRunRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - AI Vocabulary
      summary: List AI vocabulary extraction runs
      operationId: list_runs_ai_vocabulary_runs_get
      parameters:
      - name: dataset
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Dataset
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RunResponse'
                title: Response List Runs Ai Vocabulary Runs Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /ai-vocabulary/runs/{run_id}:
    get:
      tags:
      - AI Vocabulary
      summary: Get AI vocabulary extraction run
      operationId: get_run_ai_vocabulary_runs__run_id__get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /ai-vocabulary/runs/{run_id}/execute:
    post:
      tags:
      - AI Vocabulary
      summary: Execute AI vocabulary extraction run
      operationId: execute_run_ai_vocabulary_runs__run_id__execute_post
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /ai-vocabulary/runs/{run_id}/execute-async:
    post:
      tags:
      - AI Vocabulary
      summary: Queue AI vocabulary extraction run
      operationId: execute_run_async_ai_vocabulary_runs__run_id__execute_async_post
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /ai-vocabulary/runs/{run_id}/terms:
    get:
      tags:
      - AI Vocabulary
      summary: List run terms with filters
      operationId: list_run_terms_ai_vocabulary_runs__run_id__terms_get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      - name: validationStatus
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Validationstatus
      - name: confidenceMin
        in: query
        required: false
        schema:
          anyOf:
          - type: number
          - type: 'null'
          title: Confidencemin
      - name: hasCandidate
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Hascandidate
      - name: docId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Docid
      - name: term
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Term
      - name: sortBy
        in: query
        required: false
        schema:
          type: string
          default: createdAt
          title: Sortby
      - name: sortOrder
        in: query
        required: false
        schema:
          type: string
          default: desc
          title: Sortorder
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page
      - name: size
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunTermsPageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /ai-vocabulary/candidates:
    get:
      tags:
      - AI Vocabulary
      summary: List AI vocabulary candidates
      operationId: list_candidates_ai_vocabulary_candidates_get
      parameters:
      - name: dataset
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Dataset
      - name: ai_run_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Ai Run Id
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TermCandidateResponse'
                title: Response List Candidates Ai Vocabulary Candidates Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /ai-vocabulary/candidates/{candidate_id}:
    get:
      tags:
      - AI Vocabulary
      summary: Get candidate vocabulary detail
      operationId: get_candidate_ai_vocabulary_candidates__candidate_id__get
      parameters:
      - name: candidate_id
        in: path
        required: true
        schema:
          type: integer
          title: Candidate Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TermCandidateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /ai-vocabulary/candidates/{candidate_id}/evidence:
    get:
      tags:
      - AI Vocabulary
      summary: List candidate evidence rows
      operationId: list_candidate_evidence_ai_vocabulary_candidates__candidate_id__evidence_get
      parameters:
      - name: candidate_id
        in: path
        required: true
        schema:
          type: integer
          title: Candidate Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TermCandidateEvidenceResponse'
                title: Response List Candidate Evidence Ai Vocabulary Candidates  Candidate
                  Id  Evidence Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /ai-vocabulary/candidates/{candidate_id}/review:
    post:
      tags:
      - AI Vocabulary
      summary: Review candidate vocabulary term
      operationId: review_candidate_ai_vocabulary_candidates__candidate_id__review_post
      parameters:
      - name: candidate_id
        in: path
        required: true
        schema:
          type: integer
          title: Candidate Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewCandidateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TermCandidateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /ai-vocabulary/prompts:
    get:
      tags:
      - AI Vocabulary
      summary: List prompt versions
      operationId: list_prompt_versions_ai_vocabulary_prompts_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/PromptVersionResponse'
                type: array
                title: Response List Prompt Versions Ai Vocabulary Prompts Get
    post:
      tags:
      - AI Vocabulary
      summary: Create prompt version
      operationId: create_prompt_version_ai_vocabulary_prompts_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePromptVersionRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromptVersionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /ai-vocabulary/prompts/{prompt_version}:
    get:
      tags:
      - AI Vocabulary
      summary: Get prompt version
      operationId: get_prompt_version_ai_vocabulary_prompts__prompt_version__get
      parameters:
      - name: prompt_version
        in: path
        required: true
        schema:
          type: string
          title: Prompt Version
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromptVersionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /ai-vocabulary/runs/{run_id}/summary:
    get:
      tags:
      - AI Vocabulary
      summary: Get run analytics summary
      operationId: get_run_summary_ai_vocabulary_runs__run_id__summary_get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunDetailSummaryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /ai-vocabulary/runs/{run_id}/invalid-breakdown:
    get:
      tags:
      - AI Vocabulary
      summary: Get invalid term breakdown
      operationId: get_invalid_breakdown_ai_vocabulary_runs__run_id__invalid_breakdown_get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidBreakdownResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /ai-vocabulary/runs/{run_id}/top-candidates:
    get:
      tags:
      - AI Vocabulary
      summary: Get top linked candidates for a run
      operationId: get_top_candidates_ai_vocabulary_runs__run_id__top_candidates_get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunTopCandidatesResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /ai-vocabulary/runs/{run_id}/compare:
    get:
      tags:
      - AI Vocabulary
      summary: Compare two runs
      operationId: compare_runs_ai_vocabulary_runs__run_id__compare_get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      - name: targetRunId
        in: query
        required: true
        schema:
          type: string
          title: Targetrunid
      - name: topN
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Topn
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompareRunResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CreateRunRequest:
      properties:
        dataset:
          type: string
          title: Dataset
        sample_version_id:
          type: string
          title: Sample Version Id
        prompt_version:
          type: string
          title: Prompt Version
          default: vocab_extract_v1
        provider:
          type: string
          title: Provider
          default: openai
        model_name:
          type: string
          title: Model Name
          default: ''
        temperature:
          type: number
          maximum: 2.0
          minimum: 0.0
          title: Temperature
          default: 0.1
        batch_size:
          type: integer
          maximum: 50.0
          minimum: 1.0
          title: Batch Size
          default: 10
      type: object
      required:
      - dataset
      - sample_version_id
      title: CreateRunRequest
...
    GenerateSampleRequest:
      properties:
        dataset:
          type: string
          title: Dataset
          description: Dataset identifier
        version_name:
          type: string
          title: Version Name
          description: Frozen sample version name, e.g. v1
        sample_type:
          type: string
          title: Sample Type
          description: BASE | TOPIC | BLIND
          default: BASE
        candidate_pool_size:
          type: integer
          maximum: 20000.0
          minimum: 100.0
          title: Candidate Pool Size
          default: 2000
        sample_size:
          type: integer
          maximum: 1000.0
          minimum: 10.0
          title: Sample Size
          default: 150
        similarity_threshold:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Similarity Threshold
          default: 0.85
        max_chunks_per_doc:
          type: integer
          maximum: 10.0
          minimum: 1.0
          title: Max Chunks Per Doc
          default: 2
      type: object
      required:
      - dataset
      - version_name
      title: GenerateSampleRequest
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...